Consolidate resume features
