+++ src/index.ts
@@ -28,7 +28,6 @@ import { CHANNELS, initCachedChannels } from "./constants/channels";
 import { scheduleTask } from "./helpers/schedule";
 import { discordToken } from "./helpers/env";
 import { registerCommand, deployCommands } from "./helpers/deploy-commands";
-import resumeReviewPdf from "./features/resume-review";
 import troll from "./features/troll";
 
 export const bot = new discord.Client({
@@ -190,8 +189,6 @@ addHandler(
 
 const threadChannels = [CHANNELS.helpJs, CHANNELS.helpThreadsReact];
 addHandler(threadChannels, autothread);
-
-addHandler(CHANNELS.resumeReview, resumeReviewPdf);
 
 bot.on("ready", () => {
   deployCommands(bot);
